moves openDeleteSongDialog function to SidebarToolbar (preparation to remove EditSongDialog.kt)
