+++ src/main/kotlin/eu/florian_fuhrmann/musictimedtriggers/gui/views/app/sidebar/SidebarToolbar.kt
@@ -7,10 +7,12 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
+import eu.florian_fuhrmann.musictimedtriggers.gui.alerts.BasicAlert
 import eu.florian_fuhrmann.musictimedtriggers.gui.dialogs.DialogManager
 import eu.florian_fuhrmann.musictimedtriggers.gui.dialogs.addsong.AddSongDialog
-import eu.florian_fuhrmann.musictimedtriggers.gui.dialogs.editsong.openDeleteSongDialog
+import eu.florian_fuhrmann.musictimedtriggers.project.Project
 import eu.florian_fuhrmann.musictimedtriggers.project.ProjectManager
+import eu.florian_fuhrmann.musictimedtriggers.song.Song
 import org.jetbrains.jewel.foundation.theme.JewelTheme
 import org.jetbrains.jewel.ui.component.Icon
 import org.jetbrains.jewel.ui.component.IconButton
@@ -46,4 +48,26 @@ fun SidebarToolbar() {
             Icon(AllIconsKeys.General.Remove, null)
         }
     }
+}
+
+fun openDeleteSongDialog(project: Project, song: Song?) {
+    //show confirmation alert
+    BasicAlert(
+        type = BasicAlert.Type.Warning,
+        title = "Confirm Deletion",
+        buttons = {
+            CancelButton()
+            CancelButtonFocused()
+            OKButton(onClick = {
+                // close alert
+                close()
+                // delete song
+                if (song != null) {
+                    project.deleteSong(song)
+                }
+            }, label = "Confirm")
+        }
+    ) {
+        Text("Are you sure you want to delete Song ${song?.name}?")
+    }.show()
 }
